requirements.txt
langchain_openai
pydantic
python-docx
Pillow
requests
python-dotenv
uvicorn
fastapi
streamlit